Check how many bytes have been received from socket.
The recv function does not guarantee that it will return the right
number of bytes, so we have to check how many bytes were returned and
possibly call recv some more times. This patch adds a recvBytes
function that receives a given number of bytes, handling all the weird
situations that may arise, and uses it wherever appropriate.